cleans-up clueboards, readmes to lowercase
