+++ keyboards/clueboard2/Makefile
@@ -1,7 +1,5 @@
 
 
-SRC = led.c
-
 # MCU name
 #MCU = at90usb1287
 MCU = atmega32u4
@@ -67,10 +65,6 @@ MIDI_ENABLE ?= no 		# MIDI controls
 AUDIO_ENABLE ?= no
 UNICODE_ENABLE ?= no 		# Unicode
 BLUETOOTH_ENABLE ?= no # Enable Bluetooth with the Adafruit EZ-Key HID
-
-#ifdef BACKLIGHT_ENABLE
-    SRC := backlight.c $(SRC)
-#endif
 
 ifndef QUANTUM_DIR
 	include ../../Makefile
