+++ keyboards/clueboard2/clueboard2.h
@@ -1,9 +1,7 @@
 #ifndef CLUEBOARD2_H
 #define CLUEBOARD2_H
 
-#include "matrix.h"
-#include "keymap.h"
-#include <stddef.h>
+#include "quantum.h"
 
 
 /* Clueboard matrix layout
@@ -50,8 +48,5 @@
     { k80, k81, k82, k83, k84, k85, k86, KC_NO }, \
     { k90, KC_NO, k92, k93, k94, k95, k96, k97 }  \
 }
-
-void matrix_init_user(void);
-void matrix_scan_user(void);
 
 #endif
