+++ keyboards/kc60/keymaps/stanleylai/readme.md
