+++ keyboards/kc60/readme.md
